Move dashboard filter converters to staging
JIRA: RAIL-3492
